Re-enable script support (but needs unsafe-eval CSP)
This makes scripts work again, but it is not a final solution, as it only works with "unsafe-eval" enabled in the add-ons CSP, which is not allowed on ATN. But it is a start.

The this.mWindow object is no longer available. As an alternative, the this.mDetails object is provided, which is the WebExtension ComposeDetails for the current tab.

WebExtension APIs are now usable in scripts.