finalize script engine
Scripts are now executed in the scope of the content script, which is allowed on ATN. The script however does not have access to most WebExtension APIs. The important ones are made available by Quicktext via this.compose.* and this.messages.*

Scripts can also be executed remotely (ESCRIPT) by specifying an extensionId, which has the script and returns the evaluated string. This can be used if WebExtension APIs are required which are not among the standard APIs exposed by Quicktext, or if Experiment evaluation is required.